move events, reminder, task data objects out of domain layer and into data layer since they match JSON response; create a mapper function
