+++ app/src/main/java/com/example/tasky/feature_agenda/data/model/Event.kt
@@ -1,4 +1,6 @@
-package com.example.tasky.agenda_details.domain.model
+package com.example.tasky.feature_agenda.data.model
+
+import com.example.tasky.agenda_details.domain.model.Photo
 
 data class Event(
     val id: String,
